Move `fake_stripe_id` and `FAKE_STRIPE_ID` to conftest
