+++ tests/unit/conftest.py
@@ -1,6 +1,7 @@
 """pytest fixtures for the CTMS app"""
 import json
 import os.path
+from base64 import b64encode
 from datetime import datetime, timezone
 from glob import glob
 from time import mktime
@@ -49,7 +50,6 @@ from ctms.schemas import (
     StripeSubscriptionCreateSchema,
     StripeSubscriptionItemCreateSchema,
 )
-from tests.unit.sample_data import FAKE_STRIPE_ID
 
 from . import factories
 
@@ -68,6 +68,24 @@ SAMPLE_CONTACT_PARAMS = [
     ("simple_default_contact_data", {"amo"}),
     ("default_newsletter_contact_data", {"newsletters"}),
 ]
+
+
+def fake_stripe_id(prefix: str, seed: str, suffix: Optional[str] = None) -> str:
+    """Create a fake Stripe ID for testing"""
+    body = b64encode(seed.encode()).decode().replace("=", "")
+    return f"{prefix}_{body}{suffix if suffix else ''}"
+
+
+FAKE_STRIPE_ID = {
+    "Customer": fake_stripe_id("cus", "customer"),
+    "Invoice": fake_stripe_id("in", "invoice"),
+    "(Invoice) Line Item": fake_stripe_id("il", "invoice line item"),
+    "Payment Method": fake_stripe_id("pm", "payment_method"),
+    "Price": fake_stripe_id("price", "price"),
+    "Product": fake_stripe_id("prod", "product"),
+    "Subscription": fake_stripe_id("sub", "subscription"),
+    "Subscription Item": fake_stripe_id("si", "subscription_item"),
+}
 
 
 def _gather_examples(schema_class) -> dict[str, str]:
